Add ability to send WOL packets
We had to use a lower level library since we want to be able to specify
which network interface and an optional VLAN.

Not sure how to test this systematically.  Right now I've tested this
with Wireshark, and confirmed that packets are being sent out with the
correct VLAN tags.  Might need to get some physical machines to test
this out.